dns-updater/requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==2.3.2
google-api-core==2.19.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-dns==0.35.0
google-cloud-logging==3.10.0
google-cloud-secret-manager==2.20.0
google-oauth==1.0.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
